Repository, controller e service Form Of Payment

+++ src/main/kotlin/br/com/dastec/gerenciasalao/controllers/requests/services/PutServiceRequest.kt
@@ -1,6 +1,7 @@
 package br.com.dastec.gerenciasalao.controllers.requests
 
 import com.fasterxml.jackson.annotation.JsonAlias
+import java.math.BigDecimal
 
 data class PutServiceRequest(
 
@@ -10,7 +11,7 @@ data class PutServiceRequest(
     @JsonAlias("id_category")
     val idCategory: Long,
 
-    val price: Double?,
+    val price: BigDecimal?,
 
     val observation: String?
 )
