+++ src/main/kotlin/br/com/dastec/gerenciasalao/models/ServiceModel.kt
@@ -1,5 +1,6 @@
 package br.com.dastec.gerenciasalao.models
 
+import java.math.BigDecimal
 import javax.persistence.*
 
 @Entity
@@ -18,5 +19,5 @@ data class ServiceModel(
     @JoinColumn(name = "id_category", nullable = false)
     val category: CategoryModel,
 
-    val price: Double?
+    val price: BigDecimal?
 )
